backend/requirements.txt
flask==3.0.3
werkzeug==3.0.3
flask-cors==3.0.10
flask-session==0.5.0
requests==2.31.0
python-dotenv==1.0.0
pymongo==4.4.0
bcrypt==4.0.1
numpy==1.24.2
polyline==2.0.2
osmnx==1.3.0
networkx==2.8.0
pytest==7.4.0
pytest-flask==1.2.0
stadiamaps==5.0.0